requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.2
pillow==10.2.0
pytesseract==0.3.10
python-docx==1.0.0
pydantic-settings
googletrans==4.0.2
easyocr==1.7.2
torch==2.6.0
numpy==1.26.0
